L2TP: Handle multiple sessions per tunnel
Accept to create several sessions for each tunnel. Session IDs
are generated randomly and stored in a binary search tree in
their corresponding tunnel (i.e. field "sessions" of the
l2tp_conn_t structure). A new mempool is defined for session
allocations.

Generation of the session IDs is simple but quite limited: it
selects a 16 bits random ID and checks if a session with this
ID already exists. If so the allocation fails and the session
is closed. Otherwise the selected ID is used for the new
session. So tunnels that handle many sessions may reject new
ones, even if unused session IDs are available (just because
the randomly choosen ID is already in use).

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>