l2tp: Always acknowledge SCCCN messages
Tunnel establishment may not be immediately followed by the creation of
a new session. In this case there is no message to send to the peer
after receiving an SCCCN; so it won't be acknowledged and the peer will
then have to retransmit it.
This patch makes the tunnel always acknowledge SCCCN by replying with a
ZLB, thus avoiding useless retransmission from the peer.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>