l2tp: No automatic CDN before cancelling sessions
This patch removes the automatic sending of a CDN message when
cancelling creation of new sessions. This is necessary in order
to give the possibility for a tunnel to create (and possibly cancel)
new sessions on its own (e.g. for placing outgoing calls). In such
cases, sending a CDN message would not make sense if the session gets
cancelled before any creation message was sent.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>