cli: Modify regexp command handler registration
Pass a full string, instead of an array of words, to the "help" callback
of regexp command handlers. Also register these command handlers using
a regexp to apply on "help" commands, just like for exec commands. The
"help" callback will then be called only if the command matches the
"help" regexp.

As a side effect, the "help" word and its following spaces are skipped
before calling the "help" callback for both simple and regexp commands.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>