l2tp: Pass host address as sockaddr in tunnel alloc.
Pass the host address as a directly usable sockaddr_in in
l2tp_tunnel_alloc(), instead of an in_pktinfo. This will allow to
allocate tunnels without having to artificially create in_pktinfo
structures in situations where the tunnel is not created in response
to an SCCRQ.

As a side effect, this patch makes accel-ppp reply to SCCRQ using
an arbitrary source port, as specified in RFC 2661.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>