l2tp: Define function to compute Challenge Response
Create the l2tp_tunnel_genchallresp() function for building and
adding the Challenge Response AVP to messages that need it.
Support for this AVP is needed for sending SCCRP and SCCCN messages.

This patch replaces the previous Challenge Response computation by
calling the new function and prepares implementation of SCCCN message
sending.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>