l2tp: Update Ns/Nr handling
Change the meaning of the Nr field in struct l2tp_conn_t: instead of
setting it to the last received message sequence, set it with the
value of the next expected message sequence number (i.e. one above
the previous value). This simplifies processing and avoids the
l2tp_send() function to increment the L2TP packet Nr field (it's now
directly set to the correct value).
Having Nr correctly set before calling l2tp_send() is necessary for
implementing tunnel establishment requests (SCCRQ messages) since the
Nr field of this message must be 0.

This patch also replaces the formula used to check if a message Ns is
lower than a tunnel Nr.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>