l2tp: Send L2TP messages with sendto()
Send L2TP messages to the host set in the addr field of l2tp_packet_t,
instead of relying on the file descriptor's connection state. This
will allow the file descriptor to be disconnected (allowing it to
wait for a message from an arbitrary source port) while still being
able to retransmit previous L2TP messages.

This possibility is necessary for initiating tunnel establishment: the
initial SCCRQ is sent to the peer on port 1701, but the peer may reply
with an arbitrary source port. The socket thus needs to be disconnected
to be able to receive the peer's SCCRP. But in the mean time, the
initial SCCRQ may need to be retransmitted (if no answer is received
from the peer after a given amount of time). Hence the need for sending
L2TP messages over unconnected sockets.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>