l2tp: Flush ack-ed messages from retransmission queue
When several messages are sent consecutively, the peer can reply with
a single ZLB to acknowledge all messages. In this case, removing
only the first message from the retransmission queue isn't enough
(acknowledged messages left in the queue will be retransmitted).
This case can happen when replying to an outgoing call request, wich
is handled by sending two messages in a row (an OCRP and an OCCN).

This patch avoids useless retransmissions by removing all messages
acknowledged by peer from the retransmission queue.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>