l2tp: Fix early message sending during tunnel establishment
A side effect of setting peer port to zero when sending SCCRQ messages
is that the tunnel can't send any new message before receiving first
peer's answer (this doesn't apply to retransmissions as peer port is
set in the original packet). It is then impossible to send StopCCN
messages within this time frame.

This patch fixes this issue using a new tunnel flag which indicates if
tunnel's peer port has a default value or if it has been set after
reception of peer's first message.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>