triton: Fix race upon termination
The triton_terminate() function works by setting the need_close flag
of each triton context, then queues this context for execution by a
triton thread if not already running. But if the context is already
being run by a triton thread, it may not notice that its need_close
flag has been updated (this flag is only checked at the beginning
of ctx_thread()). So if no other event wakes up that context (i.e.
if ctx_thread() isn't run again), it will never terminate.

This patch moves the need_close flag check at the end of ctx_thread()
so that a triton context can take the need_close flag into account
event if it's updated while running.

The context spinlock is also used to protect the need_close flag as
it is concurrently updated by triton_terminate().

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>