l2tp: automatically delete tunnel if disconnection fails
Let l2tp_tunnel_disconnect() delete the tunnel when an error occurs.
Return an int to indicate if conn has been deleted.

Most callers are compatible with this new behaviour, either because
they never use the tunnel after disconnection or because they hold
a reference on the tunnel.

A few callers need to be adapted though:
  -l2tp_conn_close() calls l2tp_tunnel_disconnect() to properly inform
   the peer of tunnel disconnection. But it also deletes the tunnel
   immediately, no matter if StopCCN message was sent or not. Since
   l2tp_conn_close() doesn't hold a reference on the tunnel, care must
   be taken not to access the tunnel if it was deleted by
   l2tp_tunnel_disconnect().
   Since immediate deletion isn't required, the l2tp_tunnel_free() call
   is simply removed. The normal disconnection process is now followed,
   with detection and retransmission of lost messages, acknowledgment
   handling, etc.

  -situation is similar for l2tp_tunnel_timeout() and handled in the
   same way.

This patch also deletes tunnel's sessions, as well as the establishment
and hello timers. These are of no use once the tunnel enters the
disconnection process.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>