net-snmp: run snmp in "special" thread
NET-SNMP is built on top of stale file descriptor multiplexing mechanism (select)
which prevents to use descriptors with number >= 1024.
Due to accel-ppp uses a lots of descriptors NET-SNMP easy overflows this limit and fails
and even may crash whole accel-ppp process.
This patch runs snmp in special thread which does not share file descritor table
with main process, so it can use less descriptors and live in __FD_SETSIZE limit.

Signed-off-by: Dmitry Kozlov <xeb@mail.ru>