ppp: create ppp units after authentication
This avaid allocating a ppp unit when authentication failed

Split establish_ppp in two functions estabish_ppp and
connect_ppp_channel. The fist one connect the channel on an instance of
/dev/ppp, allocate channel resources and start first ppp layer.
The second functions create ppp unit and connect the channel to this
unit. It is called after authentication.
destablish_ppp is also split in two function for symmetry and
ppp_terminate is adapted to handle the case when the unit is not
created.

Signed-off-by: François Cachereul <f.cachereul@alphalink.fr>