pppd-compat: move computation of tx_bytes and rx_bytes
tx_bytes and rx_bytes are only used in the ->stop_time branch, so
let's compute them here rather than at the top of the function.

Also, let's replace '4294967296llu' by 'UINT32_MAX + 1', to make it
clear that a gigaword equals 2^32 bytes and is used to keep track of
how many times the 32 bits tx and rx counters have overflowed.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>