radius: add support for route priority (metric) in Framed-Route
Let an optional route priority (aka metric) be defined in RADIUS
Framed-Route attributes.
The priority is an integer placed at the end of the route string. This
is backward compatible with the previous format and also conforms with
the recommended format defined by RFC 2865 (although we don't allow
multiple metrics).

Framed-Route format is:
<network> [<gateway> [<priority>]]

For example, 'Framed-Route = "192.0.2.0/24 203.0.113.1 8"' will let
the following route be installed (assuming 203.0.113.1 is routed
through eth0):
$ ip route show
[...]
192.0.2.0/24 via 203.0.113.1 dev eth0 metric 8

It's possible to use the unspecified gateway (0.0.0.0) if one wants to
set a priority without specifying a gateway address.

Finally, route deletion now also takes the priority into account, in
order to avoid removing a different route accidentally.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>