auth: fix missing CHAP Success message in chap-md5
When pwdb returns PWDB_NO_IMPL, then chap_recv_response() tries to
authenticate the peer itself. If this authentication succeeds but the
session is already started (in case of re-authentication, with option
conf_interval), no CHAP Success is sent.

This patch sends the missing CHAP Success message in this case, so that
the peer knows that its response has been received and accepted, and
that no retransmission is required.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>