auth: avoid sending duplicate CHAP Failure messages in mschap-v1 and mschap-v2
When pwdb returns PWDB_NO_IMPL and chap_recv_response() performs the
authentication itself, it delegates this task to chap_check_response().
This function sends a CHAP Failure message if it can't retrieve the
password, but not in case of password mismatch. Since
chap_recv_response() already sends a CHAP Failure message on error,
the one sent by chap_check_response() is useless.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>