auth: fix re-authentication of peer in all chap modules
If the peer re-authenticates (because of option conf_interval) and pwdb
calls auth_result(), we may add the interval timer again to the session
context. This crashes accel-ppp when deleting the session, because the
interval timer is removed only once and the superfluous timers are
still running. Therefore, when removing the context, triton detects
this issue and calls abort().

To fix this, we need to detect if the session is already started and
just send a CHAP Success message in this case.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>