[vBNG] Add pppoe sessionid
Add the pppoe sessionid to the ap_session and publish it to redis.

[pppoe_session_id] Send pppoe sessionid to redis

Publish pppoe sessionid on redis if set.

[vBNG] Fix segfault in pppoe sessionid

The assignment of the pppoe sessionid leads to a segfault, because the
redis message is string, while the session id of the ap_session is int.
Fix the bug by making the sessionid in the redis message also int.