iputils: add 'src' and 'gw' parameters to iproute_del()
Rework iproute_del() to have the same parameters as iproute_add().
This will allow callers to specify more precisely the route they want
to delete.

Callers will later be converted to make use of these parameters to
ensure that the removed route precisely matches the one that was
originaly inserted.

Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>