net: new function get_ifindex
The index of a given interface is an operation that highly depends on
the network namespace we're in. This patch simply cuts out a function to
get the index for a given interface name from the session initialization
code, and expose it in the ap_net structure.

This function can then be used to refresh the index when moving
interfaces around.

Signed-off-by: Simon Chopin <s.chopin@alphalink.fr>