CMake: use 'git describe' over 'git log' to obtain version number
This is more efficient as its a Git build in function. If build from a non
tagged commit, the version number will be extended by the "git describe" way:

  $ dpkg -I accel-ppp/build/_CPack_Packages/Linux/DEB/accel-ppp.deb | grep Vers
  Version: 1.12.0-53-g7afe0f0

The version indicator shows that the last Accel-PPP release was 1.12.0 (which
was tagged) and we are now 53 commits past the 1.12.0 tag. The current commit
id is 7afe0f0 (short form).

After installation the Debian package database will show:
  $ dpkg -l | grep accel
  ii  accel-ppp      1.12.0-53-g7afe0f0 amd64   PPtP/L2TP/PPPoE/SSTP server for Linux

Also the embedded version command will show:
  $ accel-cmd -p 2005 -V
  accel-cmd 1.12.0-53-g7afe0f0

Signed-off-by: Christian Poessinger <christian@poessinger.com>