pppd_compat: mark all executable paths as const strings
Except when loading the configuration, the content of all those strings
is not supposed to change during the execution of the program. Even
then, the mutation should only occur in the triton module before
exposing those strings to consumption.

This patch changes the pppd_compat module to actually reflect this fact
by marking their configuration strings as const.

Signed-off-by: Simon Chopin <s.chopin@alphalink.fr>