pppd_compat: fix constness warnings in execve preparations
Apparently, execve() only has one level of const marker in order not to
break existing code, but it is not supposed to mutate the content, hence
the safety of casting our strings at the last moment.

The general pattern when initializing an array of const structures is to
first generate the instances as we want them, and only when they're
ready do we insert them into the array. That's the reason behind the
changes when filling the argv array, as the previous version did it in
the reverse order.

Signed-off-by: Simon Chopin <s.chopin@alphalink.fr>