pppoe: pass discovery packets using an explicit data structure
This way, the client code doesn't need to know the intrinsics of how
exactly the packet data is encoded within the payload, it can just
directly access the structure fields.

A small theoretical bonus is that it makes the code a bit more portable
as the size of an `int` isn't guaranteed to be 4 bytes on every
platform. Granted, it is 32-bit on all the platforms supported by
Debian, and the Linux kernel pretty much makes the same assumption, but
it doesn't cost anything.

v2: Fix the pointer arithmetic bug that rendered the whole patch
non-functional

Signed-off-by: Simon Chopin <s.chopin@alphalink.fr>

fixup