pppoe: disc: document who holds references to the net object
While working on another patch, I had a bug that *seemed* to be related
to a use-after-free on the net structure, leading to a faulty patch
removing the disc_stop refcount decrement, as I missed that the init_net
function initialized the refcount to 1 and not 0. The bug was in my
code.

This patch adds simple comments the reference handling to make it a bit
more obvious what is going on.

Signed-off-by: Simon Chopin <s.chopin@alphalink.fr>