[redis] Read from epoll fd to flush events
High CPU usage was observer on the redis event worker after the first
arrival of an event. We traced down the issue to the `epoll_pwait` call.
The epoll file descriptor was not flushed after wakeup resulting in the
worker not going back into the loop waiting for new events. By reading
from the file descriptor after wakeup the eventfd gets cleared.