radius: Insert Filter-Id into session context
Insert `Filter-Id` into `ap_session` struct. The `rad_proc_attrs`
function is called on *Access-Accept* and parses attributes from the
radius message. Some of the parsed attributes are already persisted in
`ap_session`. This commit writes the value of the `Filter-Id` attribute
into a newly introduced member of the `ap_session`. Persisting the value
in the session context provides access to the `Filter-Id` value for
later processing (e.g. most trition events pass a `ap_session` pointer
to the event listeners).