Extract subscriber tags from PPPoE tag
This commit is part of a workaround to get accel-ppp to work without
setting up pre-created tagged interfaces (our current alternative to not
use vlan-mon).

Whenever a PADI packet is received, we inspect the existing tags and
look for vendor specific tags that have the vendor ID equal to 0xaaaa
(VENDOR_BISDN). If the tag data is large enough to contain an integer
after the vendor ID, its value is assigned to a new entry of the new
subscriber_tags_t struct, which contains the tag value and the
corresponding subscriber MAC address.

When a new PPPoE channel is established (allocate_channel), we verify if
there are subscriber_tags_t entries containing the MAC address of the
subscriber, and use the tag value to populate the subscriber_tags field
in its ap_session. In addition, when processing timed-out PADI packets,
we also purge existing subscriber_tags_t entries that match the same
subscriber address.

Signed-off-by: Ricardo Santos <ricardo.santos@bisdn.de>